client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["Jeferson <souzajefison@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
